requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
Brlapi==0.8.3
build==1.2.1
buildozer==1.5.0
cachetools==5.4.0
certifi==2020.6.20
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
Cython==3.0.10
dbus-python==1.2.18
debugpy==1.8.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.21.2
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fqdn==1.5.1
frozenlist==1.4.1
genai==2.1.0
git-py==1.0.2
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.1
grpcio-status==1.62.2
gTTS==2.5.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==2.8
importlib-metadata==4.6.4
infi.execute==0.1.8
infi.monotonic-time==0.1.9
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keyring==23.5.0
Kivy==2.3.0
Kivy-Garden==0.1.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==8.10.0
multidict==6.0.5
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
newsapi-python==0.2.7
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.0.1
oauthlib==3.2.0
olefile==0.46
openai==0.27.10
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pitch==0.0.6
platformdirs==4.2.2
playsound==1.3.0
pluggy==1.5.0
portpicker==1.2.0
prometheus_client==0.20.0
prompt-toolkit==1.0.18
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproject_hooks==1.1.0
pyRFC3339==1.1
pytest==8.3.2
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-json-logger==2.0.7
pyttsx3==2.90
pytube==15.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.7.24
reportlab==3.6.8
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.1
rsa==4.9
SecretStorage==3.3.1
Send2Trash==1.8.3
sh==1.14.3
simplegeneric==0.8.1
six==1.12.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.4
stack-data==0.6.3
systemd-python==234
tabulate==0.9.0
terminado==0.13.3
tiktoken==0.3.3
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tornado==4.5.3
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.24.3
virtualenv==20.26.3
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
